Struct headers_relay::headers::QueuedHeaders

pub struct QueuedHeaders<P: HeadersSyncPipeline> { /* private fields */ }

Headers queue.

Implementations

impl<P: HeadersSyncPipeline> QueuedHeaders<P>

pub fn headers_in_status(&self, status: HeaderStatus) -> usize

Returns number of headers that are currently in given queue.

pub fn total_headers(&self) -> usize

Returns number of headers that are currently in the queue.

pub fn best_queued_number(&self) -> P::Number

Returns number of best block in the queue.

pub fn best_synced_number(&self) -> P::Number

Returns number of best synced block we have ever seen. It is either less than best_queued_number(), or points to last synced block if queue is empty.

pub fn status(&self, id: &HeaderIdOf<P>) -> HeaderStatus

Returns synchronization status of the header.

pub fn header(&self, status: HeaderStatus) -> Option<&QueuedHeader<P>>

Get oldest header from given queue.

pub fn headers(
&self,
status: HeaderStatus,
f: impl FnMut(&QueuedHeader<P>) -> bool
) -> Option<Vec<&QueuedHeader<P>>>

Get oldest headers from given queue until functor will return false.

pub fn header_response(&mut self, header: P::Header)

Appends new header, received from the source node, to the queue.

pub fn target_best_header_response(&mut self, id: &HeaderIdOf<P>)

Receive best header from the target node.

pub fn maybe_orphan_response(&mut self, id: &HeaderIdOf<P>, response: bool)

Receive target node response for MaybeOrphan request.

pub fn maybe_extra_response(&mut self, id: &HeaderIdOf<P>, response: bool)

Receive target node response for MaybeExtra request.

pub fn extra_response(&mut self, id: &HeaderIdOf<P>, extra: P::Extra)

Receive extra from source node.

pub fn completion_response(
&mut self,
id: &HeaderIdOf<P>,
completion: Option<P::Completion>
)

Receive completion response from source node.

pub fn headers_submitted(&mut self, ids: Vec<HeaderIdOf<P>>)

When header is submitted to target node.

pub fn header_completed(&mut self, id: &HeaderIdOf<P>)

When header completion data is sent to target node.

pub fn add_incomplete_headers(
&mut self,
make_header_incomplete: bool,
new_incomplete_headers: Vec<HeaderIdOf<P>>
)

Marks given headers incomplete.

pub fn incomplete_headers_response(&mut self, ids: HashSet<HeaderIdOf<P>>)

When incomplete headers ids are receved from target node.

pub fn requires_completion_data(&self, id: &HeaderIdOf<P>) -> bool

Returns true if given header requires completion data.

pub fn incomplete_header(&mut self) -> Option<HeaderIdOf<P>>

Returns id of the header for which we want to fetch completion data.

pub fn header_to_complete(&mut self) -> Option<(HeaderIdOf<P>, &P::Completion)>

Returns header completion data to upload to target node.

pub fn prune(&mut self, prune_border: P::Number)

Prune and never accept headers before this block.

pub fn clear(&mut self)

Forgets all ever known headers.