Module headers_relay::sync_loop

Entrypoint for running headers synchronization loop.

Traits

SourceClient

Source client trait.

SyncMaintain

Synchronization maintain procedure.

TargetClient

Target client trait.

Functions

metrics_prefix

Return prefix that will be used by default to expose Prometheus metrics of the finality proofs sync loop.

run

Run headers synchronization.