Trait headers_relay::sync_loop::TargetClient

pub trait TargetClient<P: HeadersSyncPipeline>: RelayClient {
    fn best_header_id<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<HeaderIdOf<P>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn is_known_header<'life0, 'async_trait>(
        &'life0 self,
        id: HeaderIdOf<P>
    ) -> Pin<Box<dyn Future<Output = Result<(HeaderIdOf<P>, bool), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn submit_headers<'life0, 'async_trait>(
        &'life0 self,
        headers: Vec<QueuedHeader<P>>
    ) -> Pin<Box<dyn Future<Output = SubmittedHeaders<HeaderIdOf<P>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn incomplete_headers_ids<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<HashSet<HeaderIdOf<P>>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn complete_header<'life0, 'async_trait>(
        &'life0 self,
        id: HeaderIdOf<P>,
        completion: P::Completion
    ) -> Pin<Box<dyn Future<Output = Result<HeaderIdOf<P>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn requires_extra<'life0, 'async_trait>(
        &'life0 self,
        header: QueuedHeader<P>
    ) -> Pin<Box<dyn Future<Output = Result<(HeaderIdOf<P>, bool), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Target client trait.

Required Methods

fn best_header_id<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<HeaderIdOf<P>, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Returns ID of best header known to the target node.

fn is_known_header<'life0, 'async_trait>(
&'life0 self,
id: HeaderIdOf<P>
) -> Pin<Box<dyn Future<Output = Result<(HeaderIdOf<P>, bool), Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Returns true if header is known to the target node.

fn submit_headers<'life0, 'async_trait>(
&'life0 self,
headers: Vec<QueuedHeader<P>>
) -> Pin<Box<dyn Future<Output = SubmittedHeaders<HeaderIdOf<P>, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Submit headers.

fn incomplete_headers_ids<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<HashSet<HeaderIdOf<P>>, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Returns ID of headers that require to be ‘completed’ before children can be submitted.

fn complete_header<'life0, 'async_trait>(
&'life0 self,
id: HeaderIdOf<P>,
completion: P::Completion
) -> Pin<Box<dyn Future<Output = Result<HeaderIdOf<P>, Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Submit completion data for header.

fn requires_extra<'life0, 'async_trait>(
&'life0 self,
header: QueuedHeader<P>
) -> Pin<Box<dyn Future<Output = Result<(HeaderIdOf<P>, bool), Self::Error>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Returns true if header requires extra data to be submitted.