Trait headers_relay::sync_types::HeadersSyncPipeline

pub trait HeadersSyncPipeline: 'static + Clone + Send + Sync {
    type Hash: Eq + Clone + Copy + Send + Sync + Debug + Display + Hash;
    type Number: BlockNumberBase;
    type Header: SourceHeader<Self::Hash, Self::Number>;
    type Extra: Clone + Send + Sync + PartialEq + Debug;
    type Completion: Clone + Send + Sync + Debug;

    const SOURCE_NAME: &'static str;
    const TARGET_NAME: &'static str;

    fn estimate_size(source: &QueuedHeader<Self>) -> usize;
}

Headers synchronization pipeline.

Required Associated Types

type Hash: Eq + Clone + Copy + Send + Sync + Debug + Display + Hash

Headers we’re syncing are identified by this hash.

type Number: BlockNumberBase

Headers we’re syncing are identified by this number.

type Header: SourceHeader<Self::Hash, Self::Number>

Type of header that we’re syncing.

type Extra: Clone + Send + Sync + PartialEq + Debug

Type of extra data for the header that we’re receiving from the source node:

extra data is required for some headers;

target node may answer if it’ll require extra data before header is submitted;

extra data available since the header creation time;

header and extra data are submitted in single transaction.

Example: Ethereum transactions receipts.

type Completion: Clone + Send + Sync + Debug

Type of data required to ‘complete’ header that we’re receiving from the source node:

completion data is required for some headers;

target node can’t answer if it’ll require completion data before header is accepted;

completion data may be generated after header generation;

header and completion data are submitted in separate transactions.

Example: Substrate GRANDPA justifications.

Required Associated Constants

const SOURCE_NAME: &'static str

Name of the headers source.

const TARGET_NAME: &'static str

Name of the headers target.

Required Methods

fn estimate_size(source: &QueuedHeader<Self>) -> usize

Function used to estimate size of target-encoded header.